Adjust libGCAL header ImageIO_impl.h to work with newer versions
Old versions had int as return type for the function ImageIO_seek
whereas newer versions use long. Thus, replace int by long.

+++ include/FireDeamon/CGAL/ImageIO_impl.h
@@ -241,7 +241,7 @@ char *ImageIO_gets( const _image *im, char *str, int size )
 
 
 CGAL_INLINE_FUNCTION
-int ImageIO_seek( const _image *im, long offset, int whence ) {
+long ImageIO_seek( const _image *im, long offset, int whence ) {
   switch(im->openMode) {
   case OM_CLOSE :
   default :
